Adds unique client meteo venezia type weather stations

+++ app/models/meteo_data.model.js
@@ -6,7 +6,7 @@ module.exports = (sequelize, Sequelize) => {
         wind_speed_knots: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: 0 } },
         wind_direction_deg: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: 0, max: 359 } },
         barometric_pressure_hPa: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: 0 } },
-        rain_today_mm: { type: Sequelize.INTEGER, defaultValue: null, validate: { min: 0 } },
+        rain_today_mm: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: 0 } },
         rain_rate_mmph: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: 0 } },
         temperature_cels: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: -273.2, max: 273 } },
         rel_humidity: { type: Sequelize.DECIMAL, defaultValue: null, validate: { min: 0, max: 1 } },
